refined user profile page, article list and article edit page

+++ src/pages/UserProfile.vue
@@ -10,9 +10,9 @@
         <el-divider></el-divider>
         <h3>My Articles</h3>
         <el-table :data="articles" style="width: 100%;">
-            <el-table-column prop="title" min-width="300" label="Title"></el-table-column>
-            <el-table-column prop="created_at" min-width="100" label="Create Time"></el-table-column>
-            <el-table-column prop="updated_at" min-width="100" label="Update Time"></el-table-column>
+            <el-table-column prop="title" min-width="250" label="Title"></el-table-column>
+            <el-table-column prop="created_at" min-width="130" label="Create Time"></el-table-column>
+            <el-table-column prop="updated_at" min-width="130" label="Update Time"></el-table-column>
             <el-table-column
                     fixed="right"
                     label="Operation"
@@ -24,6 +24,14 @@
                 </template>
             </el-table-column>
         </el-table>
+        <el-pagination
+                layout="prev, pager, next"
+                :current-page="this.page"
+                @current-change="this.onPageChange"
+                :page-size="this.pageSize"
+                :total="this.total"
+                style="margin-top: 20px">
+        </el-pagination>
         <el-button type="success" plain @click="this.clickNewArticle" style="margin-top: 18px">New Article</el-button>
     </el-card>
 </template>
@@ -43,7 +51,10 @@
                     title: "test-1",
                     created_at: 1,
                     updated_at: 2
-                }]
+                }],
+                page: 1,
+                pageSize: 5,
+                total: 0
             }
         },
         methods: {
@@ -58,14 +69,22 @@
             getUserProfileSuccess(res) {
                 console.log(res)
                 this.profile = res.data.user
-                this.listArticlesByUserId(res.data.id)
+                this.listArticlesOfUser()
             },
-            listArticlesByUserId(userId) {
-                // TODO: implement this method
-
+            listArticlesOfUser() {
+                let userId = this.$store.state.userId
+                let url = configJson.endpoint + '/api/v1/articles?page=' + (this.page - 1) + '&limit=' + this.pageSize
+                    + '&author_id=' + userId
+                axios.get(url)
+                    .then(this.listArticlesOfUserSuccess)
+                    .catch(function (err) {
+                        console.log(err)
+                    })
             },
-            listArticlesByUserIdSuccess(res) {
-                this.articles = res.data
+            listArticlesOfUserSuccess(res) {
+                console.log(res.data)
+                this.articles = res.data.articles
+                this.total = res.data.total
             },
             clickView(articleId) {
                 this.$router.push('/articles/' + articleId)
@@ -75,20 +94,39 @@
                 this.$router.push('/articles/edit/' + articleId)
             },
             clickDelete(articleId) {
-                axios.delete(configJson.endpoint + '/api/v1/articles/' + articleId)
-                    .then(this.deleteArticleSuccess)
-                    .catch(function (err) {
-                        console.log(err)
-                    })
+                this.$confirm('Are you sure to delete this article?', 'Warning', {
+                    confirmButtonText: 'Yes',
+                    cancelButtonText: 'No',
+                    type: 'warning'
+                }).then(() => {
+                    axios.delete(configJson.endpoint + '/api/v1/articles/' + articleId)
+                        .then(this.deleteArticleSuccess)
+                        .catch(function (err) {
+                            console.log(err)
+                        })
+                }).catch(() => {
+                    this.$notify({
+                        type: 'info',
+                        message: 'Discarded'
+                    });
+                });
+
             },
             deleteArticleSuccess(res) {
                 console.log(res)
+                this.$notify({
+                    type: 'success',
+                    message: 'Successfully deleted'
+                });
                 // refresh the articles list
-                let userId = this.$store.state.userId
-                this.listArticlesByUserId(userId)
+                this.listArticlesOfUser()
             },
             clickNewArticle() {
                 this.$router.push('/articles/edit')
+            },
+            onPageChange(page) {
+                this.page = page
+                this.listArticlesOfSection(this.activeSectionId)
             }
         },
         mounted: function () {
